Add files via upload
Project Description : The concept of our project is basically based on the problem faced by the security guards of the Campus in recognizing the vehicles at the enterance gate ,wheather they are college students or outsiders. In year 2023 , semester Jan-May college faculty find a solution of this problem . They give students a form to fill which includes basic information about pupils and their vehicle details like registration number and ask for a photocopy of RC of vehicle. Then after collection data the college faculty created a 10/10 inches sticker which includes a logo of Campus and name of Campus , registration number of vehicle and a unique ID . All the students have to paste the sticker on their vehicles which they refuse to , because no wants to paste that sticker on their vehicle because it looks odd . So seeing that problem we decided to make a platform that solves the problem . Our project is basically a combination of MERN stack and python (ML). Through the webapp we give allow students to register on the platfrom and fill their personal as well ass vehicle details . From the python project we use OPENCV through which we use the camera to recognise number plate of the vehicle and matches with database wheather it is a college student or not.